+++ BTDB/IOC/DependencyAttribute.cs
@@ -1,0 +1,15 @@
+using System;
+
+namespace BTDB.IOC
+{
+    [AttributeUsage(AttributeTargets.Property)]
+    public class DependencyAttribute : Attribute
+    {
+        public readonly string? Name;
+
+        public DependencyAttribute(string? name = null)
+        {
+            Name = name;
+        }
+    }
+}