+++ BTDB/IOC/INeed.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Reflection;
 
 namespace BTDB.IOC
 {
@@ -7,8 +8,9 @@ namespace BTDB.IOC
         NeedKind Kind { get; }
         Type ClrType { get; }
         bool Optional { get; }
-        object OptionalValue { get; }
+        object? OptionalValue { get; }
         bool ForcedKey { get; }
-        object Key { get; }
+        object? Key { get; }
+        PropertyInfo? PropertyInfo { get; }
     }
 }